refactor: rewrite (*s)[idx]=val via index_mut in ReborrowVisitor
Remove the Seq-specific assign_to_slice_index from basic_block.rs and
the corresponding swap_mut_current from env.rs. Instead, ReborrowVisitor
now handles (*s)[idx]=val by:
  1. reborrowing *s → s1 (updating s's current, as for any &mut write)
  2. constructing a func operand for <[T] as IndexMut<usize>>::index_mut
  3. calling the existing type_call to apply the extern spec
  4. rewriting the place to *ptr so super_assign handles the rest

All Seq-model knowledge stays in the _extern_spec_slice_index_mut spec.
Also revert accidental .gitignore modification.
