+++ .gitignore
@@ -1,3 +1,1 @@
 target/
-librust_out.rlib
-rust_out
